// src/app/models/graphuserview.ts
export interface userview{
    totalActiveMinutes: any;
    totalInactiveMinutes: any;
    miles: any;
    moves: any;
    totalCount: any;
    last_update: any;
    monthData:monthData;
    weekData:any;
    error:any;
    message:any;
    data:any;
    dailyData: any;
}

export interface monthData{
    data:any;
}